[MIN-50] Took out two tests to the misc file
